load orm models when initializing registry
