+++ packages/registry/index.js
@@ -18,5 +18,6 @@ export const {
   ROOT_MODULE, // root module ('startupjs' framework itself)
   getModule,
   createPlugin,
-  createModule
+  createModule,
+  initRegistry
 } = createRegistry({ RegistryClass, rootModuleName: ROOT_MODULE_NAME })
